requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
watchdog>=3.0.0
psutil>=5.9.0
rich>=13.0.0
colorama>=0.4.6
python-dotenv>=1.0.0
urllib3>=2.0.0
PySastrawi>=1.2.0
google-genai>=1.0.0
anthropic>=0.40.0
scrapy>=2.11.0
twisted>=23.0.0
cryptography>=42.0.0
